Add files via upload
This is a code that I wrote on Pyzo in Pyhton. It was a 8 hours project. The aim is to display the power consumed depending on the number of cycle the user wants to display. 

The challenge for this project was to learn to use the Tkinter library in less a few hours to be able to use it to create an interface.